zalupaspb/bot/requirements.txt
discord.py==2.3.2
aiohttp==3.8.6
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25 